refactor: extract purchase info
